Set GPIO when capture starts
Doing this is not as simple as just setting a GPIO on open(). To do
this precisely we need to set the GPIO when the FIFO starts running
and also do this in the RISC DMA processor to avoid scheduling jitter
of the kernel. This requires modifying things such that the RISC DMA
processor is started on open(), instead of kept running while the
driver is loaded. That has energy benefits on its own and it's the way
the official driver works anyway.